feat: combined compute test instead of a seperate file
