+++ pkg/cli/store.go
@@ -96,7 +96,7 @@ func init() {
 	set.String("certifier-latency", "", "sets artificial latency on the certifier. Defaults to empty string (not enabled) but can set m, h, s...etc")
 
 	// scorecard certifier
-	set.Bool("compute", false, "only compute the scorecard locally with the scorecard package, skipping the scorecard API. Requires GITHUB_AUTH_TOKEN")
+	set.Bool("compute", false, "only compute the scorecard with the scorecard package, skipping the scorecard API. Requires GITHUB_AUTH_TOKEN")
 
 	// deps.dev
 	// add artificial latency to throttle deps.dev
